+++ test/types.d.ts
@@ -1,0 +1,1 @@
+declare const __PLUGIN_VERSION__: string;